+++ app.js
@@ -1,80 +1,4 @@
-const firstRow = [
-  { Backquote: { id: 'Backquote', name: '`' } },
-  { Digit1: { id: 'Digit1', name: '1' } },
-  { Digit2: { id: 'Digit2', name: '2' } },
-  { Digit3: { id: 'Digit3', name: '3' } },
-  { Digit4: { id: 'Digit4', name: '4' } },
-  { Digit5: { id: 'Digit5', name: '5' } },
-  { Digit6: { id: 'Digit6', name: '6' } },
-  { Digit7: { id: 'Digit7', name: '7' } },
-  { Digit8: { id: 'Digit8', name: '8' } },
-  { Digit9: { id: 'Digit9', name: '9' } },
-  { Digit0: { id: 'Digit0', name: '0' } },
-  { Minus: { id: 'Minus', name: '-' } },
-  { Equal: { id: 'Equal', name: '=' } },
-  { Backspace: { id: 'Backspace', name: 'delete' } },
-];
-
-const secondRow = [
-  { Tab: { id: 'Tab', name: 'tab' } },
-  { q: { id: 'q', name: 'Q' } },
-  { w: { id: 'w', name: 'W' } },
-  { e: { id: 'e', name: 'E' } },
-  { r: { id: 'r', name: 'R' } },
-  { t: { id: 't', name: 'T' } },
-  { y: { id: 'y', name: 'Y' } },
-  { u: { id: 'u', name: 'U' } },
-  { i: { id: 'i', name: 'I' } },
-  { o: { id: 'o', name: 'O' } },
-  { p: { id: 'p', name: 'P' } },
-  { BracketLeft: { id: 'BracketLeft', name: '[' } },
-  { BracketRight: { id: 'BracketRight', name: ']' } },
-  { Backslash: { id: 'Backslash', name: '\\' } },
-];
-
-const thirdRow = [
-  { CapsLock: { id: 'CapsLock', name: 'caps lock' } },
-  { a: { id: 'a', name: 'A' } },
-  { s: { id: 's', name: 'S' } },
-  { d: { id: 'd', name: 'D' } },
-  { f: { id: 'f', name: 'F' } },
-  { g: { id: 'g', name: 'G' } },
-  { h: { id: 'h', name: 'H' } },
-  { j: { id: 'j', name: 'J' } },
-  { k: { id: 'k', name: 'K' } },
-  { l: { id: 'l', name: 'L' } },
-  { Semicolon: { id: 'Semicolon', name: ';' } },
-  { Quote: { id: 'Quote', name: '\'' } },
-  { Enter: { id: 'Enter', name: 'enter' } },
-];
-
-const fourthRow = [
-  { ShiftLeft: { id: 'ShiftLeft', name: 'shift' } },
-  { z: { id: 'z', name: 'Z' } },
-  { x: { id: 'x', name: 'X' } },
-  { c: { id: 'c', name: 'C' } },
-  { v: { id: 'v', name: 'V' } },
-  { b: { id: 'b', name: 'B' } },
-  { n: { id: 'n', name: 'N' } },
-  { m: { id: 'm', name: 'M' } },
-  { Comma: { id: 'Comma', name: ',' } },
-  { Period: { id: 'Period', name: '.' } },
-  { Slash: { id: 'Slash', name: '/' } },
-  { ArrowUp: { id: 'ArrowUp', name: '▲' } },
-  { ShiftRight: { id: 'ShiftRight', name: 'shift' } },
-];
-
-const fifthRow = [
-  { ControlLeft: { id: 'ControlLeft', name: 'control' } },
-  { AltLeft: { id: 'AltLeft', name: 'alt' } },
-  { MetaLeft: { id: 'MetaLeft', name: 'command' } },
-  { Space: { id: 'Space', name: ' ' } },
-  { MetaRight: { id: 'MetaRight', name: 'command' } },
-  { AltRight: { id: 'AltRight', name: 'alt' } },
-  { ArrowLeft: { id: 'ArrowLeft', name: '◀' } },
-  { ArrowDown: { id: 'ArrowDown', name: '▼' } },
-  { ArrowRight: { id: 'ArrowRight', name: '▶' } },
-];
+import Keyboard from './keyboard.js';
 
 const page = document.getElementById('page');
 
@@ -82,22 +6,5 @@ const app = document.createElement('div');
 app.className = 'app';
 page.append(app);
 
-const createKeysRow = (row) => {
-  const keysRow = document.createElement('div');
-  keysRow.className = 'row';
-
-  for (let i = 0; i < row.length; i++) {
-    const key = document.createElement('button');
-    key.className = 'key';
-    key.textContent = Object.values(row[i])[0].name;
-    keysRow.append(key);
-  }
-
-  app.append(keysRow);
-};
-
-createKeysRow(firstRow);
-createKeysRow(secondRow);
-createKeysRow(thirdRow);
-createKeysRow(fourthRow);
-createKeysRow(fifthRow);
+const keyboard = new Keyboard();
+app.append(keyboard.render());
